Api only works on the beta
I added a message letting people know that for now the api only works on the beta website. I had tried to use it on the normal website with no luck and it was not so clear that currently it only works on the beta website if you are only reading the docs. I assume it will at some point get moved over to the main url in which case this message can probably be removed.